// client/src/components/specific/Admin/Celeb/interface.ts
import { ReactChild } from "react";

export interface AdminCelebFormProps {
  celebKorNameField: ReactChild;
  celebEngNameField: ReactChild;
  celebRealNameField: ReactChild;
  celebCompanyField: ReactChild;
  celebGroupField: ReactChild;
  celebBirthDateField: ReactChild;
  celebDebutDateField: ReactChild;
  celebGenderField: ReactChild;
  celebRolesField: ReactChild;
  celebProfileImageField: ReactChild;
  celebDescriptionField: ReactChild;
  celebNationalityField: ReactChild;
}